Remove dynamic watchdog
JS code have a watchdog timer that will check if last data from unsee isn't too old, if that happens it will print an error in the UI and reload the page. This is a fail safe for weird bugs and cases where unsee is running unattended. Currently timeout for this watchdog is configured dynamically based on refresh rate, but if the backend is configured to pull data from Alertmanager less often than watchdog tolerance, then watchdog will trigger false positive. This commit changes watchdog tolerance to 15 minutes (it seems reasonable to assume people won't use Alertmanager pull interval longer than 1-5 minutes) and adds a note to the readme.