feat(backend): Add option to order annotations
The `annotations.order` option lets users pass a list of annotation names in
the order they want them displayed. Any annotation that isn't in the list is
shown after the annotations from the list, in alphabetical order.

For instance, with

```yaml
annotations:
  order:
  - summary
  - description
```

an alert with `description`, `help`, `runbook` and `summary` annotations will
have them displayed in that order:

- `summary`
- `description`
- `help`
- `runbook`

Signed-off-by: Benoît Knecht <bknecht@protonmail.ch>