fix(ui): Reverse ToggleIcon
The ToggleIcon works counterintuitive to how you would expect. The chevron direction should indicate what the button SHOULD do, rather that the present state. Currently when the details are hidden, the chevron points up, as if the details view is already seen, and you can hide it, rather than expand it.

The change reverses the chevron direction, so its more intuitive.